cmd/get: fix bug on copying bad file over SFTP.
SFTPCopyFileToRemote was opening file without io.O_TRUNC
which makes copy only write bytes over the file till the length
of the source file, if the remote file exists then the trailing bits
are not dealt with. This makes bad files on the remote server. Fix that
by opening all remote fils with os.O_TRUNC

Signed-off-by: Divya Antony J R <antonyjr@pm.me>